requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
gunicorn==20.1.0
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
whitenoise==6.2.0
